digital-diagnosis-mvp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pandas==2.1.3
google-generativeai==0.3.2
pydantic[email]==2.5.1
jinja2==3.1.2
aiofiles==23.2.1
python-dateutil==2.8.2